cmd/netsetgo: Enforce pid argument
If we call netsetgo without arguments, it returns an error:
ERROR - no such process

It fails because Veth.MoveToNetworkNamespace calls
netlink.LinkSetNsPid, which needs a pid that belongs to another netns in
order to set a veth interface inside the new netns.

So, before doing any setup, check if a pid was passed as argument of
netsetgo.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>